refactor: react-icons 최적화
